wrap RemoveUserRole migration in a conditional statement to avoid migration errors

+++ db/migrate/20110914203809_remove_user_role_from_users.rb
@@ -1,6 +1,8 @@
 class RemoveUserRoleFromUsers < ActiveRecord::Migration
   def up
-    remove_column  :users, :user_role_id
+    if User.column_names.include? "user_role_id"
+      remove_column  :users, :user_role_id
+    end
   end
 
   def down
